fix: improve error messages during encryption
The aim was that in addition to saying what the state is, there should
be some additional guidance on a direction to fix the error.

[#181481018](https://www.pivotaltracker.com/story/show/181481018)

+++ cmd/serve.go
@@ -146,8 +146,11 @@ func setupDBEncryption(db *gorm.DB, logger lager.Logger) storage.Encryptor {
 		}
 	}
 
-	if err := encryption.DeletePasswordMetadata(db, config.ToDeleteLabels); err != nil {
-		logger.Fatal("Error deleting stale password metadata", err)
+	if len(config.ToDeleteLabels) > 0 {
+		logger.Info("removing-state-password-metadata", lager.Data{"labels": config.ToDeleteLabels})
+		if err := encryption.DeletePasswordMetadata(db, config.ToDeleteLabels); err != nil {
+			logger.Fatal("Error deleting stale password metadata", err)
+		}
 	}
 
 	logger.Info("database-encryption", lager.Data{"primary": labelName(config.ConfiguredPrimaryLabel)})
